requirements.txt
Django>=3.2
django-cms>=3.7.4,<4
beautifulsoup4>=4.9.3
requests>=2.25.1
